Separate wallet and account
