+++ ain/wallet.py
@@ -1,4 +1,5 @@
-from typing import TYPE_CHECKING, Any, Dict, Optional, Union
+from typing import TYPE_CHECKING, Any, Optional
+from ain.account import Account, Accounts
 from ain.types import (
     TransactionBody,
     ValueOnlyTransactionInput,
@@ -9,9 +10,6 @@ from ain.utils import (
     toChecksumAddress,
     bytesToHex,
     pubToAddress,
-    privateToPublic,
-    privateToAddress,
-    mnemonicToPrivatekey,
     ecSignMessage,
     ecSignTransaction,
     ecRecoverPub,
@@ -20,43 +18,6 @@ from ain.utils import (
 
 if TYPE_CHECKING:
     from ain.ain import Ain
-
-class Account:
-    private_key: str
-    public_key: str
-    address: str
-    
-    def __init__(self, privateKey: Union[bytes, str] = None):
-        if type(privateKey) is bytes:
-            privateKeyBytes = privateKey
-        elif type(privateKey) is str:
-            privateKeyBytes = bytes.fromhex(privateKey)
-        else:
-            raise TypeError('privateKey has invalid type')
-
-        self.private_key = privateKeyBytes.hex()
-        self.public_key = privateToPublic(privateKeyBytes).hex()
-        self.address = privateToAddress(privateKeyBytes)
-
-    @classmethod
-    def fromMnemonic(cls, mnemonic: str, index: int = 0):
-        """
-        Returns an Account with the given mnemonic.
-        """
-        return cls(mnemonicToPrivatekey(mnemonic, index))
-
-    def __str__(self):
-        return "\n".join(
-            (
-                "{",
-                f"  address: '{self.address}'",
-                f"  private: '{self.private_key}'",
-                f"  public: '{self.public_key}'",
-                "}",
-            )
-        )
-
-Accounts = Dict[str, Account]
 
 class Wallet:
     defaultAccount: Optional[Account]
